requirements.txt
torch
torchvision
Pillow
tqdm
PyYAML
